requirements.txt
tensorflow>=2.13
opencv-python
numpy
matplotlib
scikit-learn
pillow
